Split between synchronous and asynchronous channels.
